+++ scripts/deploy-stage0-auction.ts
@@ -4,16 +4,22 @@ import hre from 'hardhat';
 import {deployUmbMultiSig, multiSigContract} from "./deployers/UmbMultiSig";
 import {deployLibStrings} from "./deployers/LibStrings";
 import {deployAuction} from "./deployers/Auction";
-import {deployUMB, UmbContract} from "./deployers/UMB";
-import {getDefaultOwnerBasedOnConfig, getProvider, validationMark, waitForTx} from "./helpers";
+import {deployUMB} from "./deployers/UMB";
+import {
+  checkTxSubmission,
+  currentTimestamp,
+  getProvider,
+  validationMark,
+  waitForTx, wasTxExecuted
+} from "./helpers";
 
 const {ethers, getNamedAccounts} = hre;
 
 async function main() {
   const {deployer, multiSigOwner1} = await getNamedAccounts();
+  console.log(`\n\n${'-'.repeat(80)}\nSCRIPT STARTS AT: ${currentTimestamp}\n\n`);
   console.log('ENV:', CONFIG.env);
-  console.log('DEPLOYING FROM ADDRESS:', deployer);
-
+  console.log('DEPLOYING FROM ADDRESS:', deployer, `\n\n`);
 
   if (CONFIG.auction.address) {
     console.log(validationMark())
@@ -30,77 +36,55 @@ async function main() {
   }
 
   const provider = getProvider()
-  let owner;
   let multiSig
 
-  if (CONFIG.multiSig.useIt) {
-    console.log('MULTISIG IN USE');
-    if (!CONFIG.multiSig.address) {
-      const libStrings = await deployLibStrings();
-      multiSig = await deployUmbMultiSig(libStrings.address);
-      owner = multiSig.address
-    } else {
-      console.log('MultiSig address is already set');
-      owner = CONFIG.multiSig.address;
-      multiSig = await multiSigContract()
-    }
+  if (!CONFIG.multiSig.address) {
+    const libStrings = CONFIG.libs.strings || (await deployLibStrings()).address;
+    multiSig = await deployUmbMultiSig(libStrings);
   } else {
-    console.log('MULTISIG IS OFF');
-    owner = await getDefaultOwnerBasedOnConfig();
+    console.log('MultiSig address is already set');
+    multiSig = await multiSigContract()
   }
 
-  if (!owner) {
-    console.log(validationMark(false))
-    throw Error('owner is missing')
-  }
-
-  console.log('OWNER for contracts:', owner);
-
-  const umb = await deployUMB(owner)
-  const auction = await deployAuction(umb.address, owner);
+  const umb = await deployUMB(multiSig.address)
+  const auction = await deployAuction(multiSig.address, umb.address);
 
   console.log('TOKEN for auction:', umb.address);
 
-  if (CONFIG.auction.tokensForAuction) {
-    if (CONFIG.multiSig.useIt) {
-      if (!multiSig) {
-        console.log(validationMark())
-        throw Error('multiSig contract is not set')
-      }
-
-      const multiSigOwner1Wallet = (await ethers.getSigners())[<number>hre.config.namedAccounts.multiSigOwner1];
-
-      if ((await multiSig.ownersPowers(multiSigOwner1)).toString() == '0') {
-        console.log(validationMark())
-        console.log('Owner for MultiSig is not set or it is invalid, so you need to mint tokens for Auction via Etherscan')
-        return
-      }
-
-      console.log('fire up tx for mint tokens for auction,',
-        'it will succeed only if we have enough power, otherwise other owners needs to confirm');
-
-      // mint tokens for auction
-      const tx = await multiSig.connect(multiSigOwner1Wallet).submitTokenMintTx(umb.address, auction.address, CONFIG.auction.tokensForAuction)
-      await waitForTx(tx.hash, provider)
-
-      console.log('Tx need additional owners to confirm:', !(await multiSig.isExceuted(0)))
-      console.log('Tx details:', await multiSig.getTransaction(0))
-    } else {
-      // mint tokens for auction
-      const [deployerWallet] = await ethers.getSigners();
-
-      const tx = await umb.connect(deployerWallet).mint(auction.address, CONFIG.auction.tokensForAuction)
-      await waitForTx(tx.hash, provider)
+  if (CONFIG.auction.amountOfTokensForAuction) {
+    if (!multiSig) {
+      console.log(validationMark())
+      throw Error('multiSig contract is not set')
     }
+
+    const multiSigOwnerWallet = (await ethers.getSigners())[<number>hre.config.namedAccounts.multiSigOwner1];
+
+    if ((await multiSig.ownersPowers(multiSigOwner1)).toString() == '0') {
+      console.log(validationMark())
+      console.log('Owner for MultiSig is not set or it is invalid, so you need to mint tokens for Auction via Etherscan')
+      return
+    }
+
+    console.log(`\nMinting tokens for auction...\n`);
+
+    const tx = await multiSig
+      .connect(multiSigOwnerWallet)
+      .submitTokenMintTx(umb.address, auction.address, CONFIG.auction.amountOfTokensForAuction)
+
+    let txId = checkTxSubmission(multiSig, await waitForTx(tx.hash, provider));
+    await wasTxExecuted(multiSig, txId)
 
     console.log('Balance of auction contract:', (await umb.balanceOf(auction.address)).toString())
   } else {
-    console.log('`tokensForAuction` is empty, so script did not mint tokens for auction - do it via Etherscan');
+    console.log('`CONFIG.auction.amountOfTokensForAuction` is empty, so script did not mint tokens for auction - do it via Etherscan');
   }
 }
 
 main()
-  .then(() => process.exit(0))
+  .then(() => {
+    console.log(`\n\nDONE.\n${'='.repeat(80)}\n`);
+    process.exit(0)
+  })
   .catch(error => {
     console.error(error);
     process.exit(1);