+++ scripts/deployers/Auction.ts
@@ -4,15 +4,15 @@ import hre from 'hardhat';
 
 const { ethers } = hre;
 
-export const deployAuction = async (token: string | undefined = CONFIG.UMB.address, owner: string) => {
+export const deployAuction = async (owner: string, token: string | undefined = CONFIG.UMB.address) => {
   const Contract = await ethers.getContractFactory('Auction');
-  const contract = await Contract.deploy(token, owner);
+  const contract = await Contract.deploy(owner, token);
   await contract.deployed();
 
   console.log('Auction:', contract.address);
 
   const constructorTypes = ['address', 'address']
-  const constructorArgs = [token, owner]
+  const constructorArgs = [owner, token]
 
   console.log('constructor abi, use it to validate contract:')
   console.log(constructorAbi(constructorTypes, constructorArgs))