+++ scripts/deployers/LibStrings.ts
@@ -7,6 +7,6 @@ export const deployLibStrings = async () => {
   const contract = await Contract.deploy();
   await contract.deployed();
 
-  console.log('Strings:', contract.address);
+  console.log('deployed Strings:', contract.address);
   return contract;
 };