+++ scripts/deployers/Rewards.ts
@@ -7,6 +7,6 @@ export const deployRewards = async (owner: string) => {
   const rewards = await Rewards.deploy(owner);
   await rewards.deployed();
 
-  console.log('Rewards:', rewards.address);
+  console.log('deployed Rewards:', rewards.address);
   return rewards;
 };