+++ scripts/deployers/rUMB.ts
@@ -3,20 +3,29 @@ import hre from 'hardhat';
 
 const { ethers } = hre;
 
-export const deployUMB = async (owner: string) => {
-  const Contract = await ethers.getContractFactory('rUMB');
-  const {initialHolder, initialBalance, maxAllowedTotalSupply, name, symbol} = CONFIG.UMB
+export const deployRUMB1 = async (owner: string, libStrings: string) => {
+  const Contract = await ethers.getContractFactory('rUMB', {
+    libraries: {
+      Strings: libStrings
+    }
+  });
+
+  const {stage1} = CONFIG
 
   const contract = await Contract.deploy(
-    initialHolder,
-    initialBalance,
     owner,
-    maxAllowedTotalSupply,
-    name,
-    symbol
+    stage1.rUmb.initialHolder,
+    stage1.rUmb.initialBalance,
+    stage1.rUmb.maxAllowedTotalSupply,
+    stage1.rUmb.rewardId,
+    stage1.rUmb.swapDuration,
+    CONFIG.UMB.name,
+    CONFIG.UMB.symbol,
   );
 
   await contract.deployed();
-  console.log('UMB:', contract.address);
+  console.log('Reward Token:', contract.address);
+  console.log('________name:', await contract.name());
+  console.log('______symbol:', await contract.symbol());
   return contract;
 };